Remove OGRS3 & OGP from risk response & update all associated tests

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/arnsriskactuarialapi/dto/api/PredictorResponse.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.arnsriskactuarialapi.dto.api
 
 import uk.gov.justice.digital.hmpps.arnsriskactuarialapi.dto.ValidationError
+import java.math.BigDecimal
 
 data class PredictorResponse<T : PredictorOutputResponse>(
   val algorithm: AlgorithmResponse,
@@ -8,6 +9,6 @@ data class PredictorResponse<T : PredictorOutputResponse>(
   val modelVersion: String = "1.0",
   val thresholdsVersion: String = "1.0",
   val output: T,
-  val featureValues: Map<String, Double> = emptyMap(),
+  val featureValues: Map<String, BigDecimal> = emptyMap(),
   val validationErrors: List<ValidationError>,
 )
